nearing mvp

+++ src/main/java/bankingapp/daos/SessionDAO.java
@@ -4,7 +4,7 @@ import java.util.ArrayList;
 
 public interface SessionDAO extends BasicDAO{
 	
-	public void applyForAccount(String string, double parseDouble);
+	public void applyForAccount(String string, float parseDouble);
 	public ArrayList<AccountsDAO> getPendingAccounts();
 	
 	public ArrayList<AccountsDAO> getActiveAccounts();
