+++ src/main/java/bankingapp/pojos/TransferPOJO.java
@@ -4,7 +4,7 @@ public class TransferPOJO {
 
 	protected int fromuid=0;
 	protected int fromaid=0;
-	protected double amount=0;
+	protected float amount=0;
 	protected int touid=0;
 	protected int toaid=0;
 	protected int transferID=-1;
@@ -15,7 +15,7 @@ public class TransferPOJO {
 		this.transferID = transferID;
 	}
 	public TransferPOJO() {};
-	public TransferPOJO(int transferID,int fromuid, int fromaid, double amount, int toaid) {
+	public TransferPOJO(int transferID,int fromuid, int fromaid, float amount, int toaid) {
 		super();
 		this.transferID = transferID;
 		this.fromuid = fromuid;
@@ -35,10 +35,10 @@ public class TransferPOJO {
 	public void setFromaid(int fromaid) {
 		this.fromaid = fromaid;
 	}
-	public double getAmount() {
+	public float getAmount() {
 		return amount;
 	}
-	public void setAmount(double amount) {
+	public void setAmount(float amount) {
 		this.amount = amount;
 	}
 	public int getTouid() {
